Refactor BottomNavigationBar
